--- src/implementation/make/maker/lib/check/Step.ts
@@ -1,3 +1,0 @@
-import { TestConfiguration } from './TestConfiguration.ts'
-
-export type Step = (configuration: TestConfiguration) => TestConfiguration
