+++ src/implementation/make/maker/lib/test/TestConfiguration.ts
@@ -1,6 +1,6 @@
 import { Buffer } from 'buffer'
 
-import { Maker } from '../../../maker/maker.ts'
+import { Maker } from '../../maker.ts'
 import { Code } from '../../../../../core/code/code.ts'
 
 export interface TestConfiguration {
