+++ src/implementation/make/maker/lib/test/TestStep.ts
@@ -1,0 +1,3 @@
+import { TestConfiguration } from './TestConfiguration.ts'
+
+export type TestStep = (configuration: TestConfiguration) => TestConfiguration
