+++ src/implementation/make/maker/makers/basicOne/makeEd25519Seed.ts
@@ -1,7 +1,7 @@
 import { Maker } from '../../maker.ts'
 import { Ed25519Seed } from '../../../../../core/primitive/primitives.ts'
-import { primitiveIs32Bytes } from '../../lib/validation/validations/validateLength.ts'
-import { makeSureThat } from '../../lib/validation/validate.ts'
+import { primitiveIs32Bytes } from '../../lib/validate/validations/lengthIs.ts'
+import { makeSureThat } from '../../lib/validate/validate.ts'
 
 export const makeEd25519Seed: Maker = (primitive: Buffer): Ed25519Seed => {
   makeSureThat(
