+++ src/implementation/make/maker/makers/basicOne/makeShort.ts
@@ -1,8 +1,8 @@
 import { Buffer } from 'buffer'
 import { Maker } from '../../maker.ts'
 import { Short } from '../../../../../core/primitive/primitives.ts'
-import { makeSureThat } from '../../lib/validation/validate.ts'
-import { primitiveIsTwoBytes } from '../../lib/validation/validations/validateLength.ts'
+import { makeSureThat } from '../../lib/validate/validate.ts'
+import { primitiveIsTwoBytes } from '../../lib/validate/validations/lengthIs.ts'
 
 // 2-byte base-2 number
 export const makeShort: Maker = (raw: Buffer): Short => {
