+++ src/main.ts
@@ -1,5 +1,5 @@
-import { makeSureThat } from './implementation/make/maker/lib/validation/validate.ts'
-import { primitiveIsEightBytes } from './implementation/make/maker/lib/validation/validations/validateLength.ts'
+import { makeSureThat } from './implementation/make/maker/lib/validate/validate.ts'
+import { primitiveIsEightBytes } from './implementation/make/maker/lib/validate/validations/lengthIs.ts'
 
 const raw = Buffer.from([1, 2, 3, 4, 5, 6, 7, 8])
 const rawIsCorrectLength = primitiveIsEightBytes(raw)
